Allow lazy type loader to return null
