+++ tests/Type/LazyTypeLoaderTest.php
@@ -4,30 +4,18 @@ declare(strict_types=1);
 
 namespace GraphQL\Tests\Type;
 
-use DMS\PHPUnitExtensions\ArraySubset\ArraySubsetAsserts;
-use Exception;
 use GraphQL\Error\InvariantViolation;
 use GraphQL\Type\Definition\InputObjectType;
 use GraphQL\Type\Definition\InterfaceType;
 use GraphQL\Type\Definition\ObjectType;
 use GraphQL\Type\Definition\Type;
 use GraphQL\Type\Schema;
-use PHPUnit\Framework\TestCase;
-use stdClass;
-use Throwable;
-use TypeError;
 
 /**
  * @see TypeLoaderTest
  */
-final class LazyTypeLoaderTest extends TestCase
+final class LazyTypeLoaderTest extends TypeLoaderTest
 {
-    use ArraySubsetAsserts;
-
-    private ObjectType $query;
-
-    private ObjectType $mutation;
-
     /** @var callable */
     private $node;
 
@@ -43,18 +31,12 @@ final class LazyTypeLoaderTest extends TestCase
     /** @var callable */
     private $postStoryMutationInput;
 
-    /** @var callable */
-    private $typeLoader;
-
-    /** @var array<int, string> */
-    private array $calls;
-
     /** @var array<string, Type> */
     private array $loadedTypes = [];
 
     public function setUp(): void
     {
-        $this->calls = [];
+        parent::setUp();
 
         $this->node                   = $this->lazyLoad('Node');
         $this->blogStory              = $this->lazyLoad('BlogStory');
@@ -197,33 +179,6 @@ final class LazyTypeLoaderTest extends TestCase
         };
     }
 
-    public function testSchemaAcceptsTypeLoader(): void
-    {
-        $this->expectNotToPerformAssertions();
-        new Schema([
-            'query'      => new ObjectType([
-                'name'   => 'Query',
-                'fields' => ['a' => Type::string()],
-            ]),
-            'typeLoader' => static function (): void {
-            },
-        ]);
-    }
-
-    public function testSchemaRejectsNonCallableTypeLoader(): void
-    {
-        $this->expectException(TypeError::class);
-        $this->expectExceptionMessageMatches('/callable.*, array given/');
-
-        new Schema([
-            'query'      => new ObjectType([
-                'name'   => 'Query',
-                'fields' => ['a' => Type::string()],
-            ]),
-            'typeLoader' => [],
-        ]);
-    }
-
     public function testWorksWithoutTypeLoader(): void
     {
         $schema = new Schema([
@@ -300,108 +255,17 @@ final class LazyTypeLoaderTest extends TestCase
         );
     }
 
-    public function testOnlyCallsLoaderOnce(): void
-    {
-        $schema = new Schema([
-            'query'      => $this->query,
-            'typeLoader' => $this->typeLoader,
-        ]);
-
-        $schema->getType('Node');
-        self::assertEquals(['Node'], $this->calls);
-
-        $schema->getType('Node');
-        self::assertEquals(['Node'], $this->calls);
-    }
-
-    public function testFailsOnNonExistentType(): void
-    {
-        $schema = new Schema([
-            'query'      => $this->query,
-            'typeLoader' => static function (): void {
-            },
-        ]);
-
-        $this->expectException(InvariantViolation::class);
-        $this->expectExceptionMessage('Type loader is expected to return a callable or valid type "NonExistingType", but it returned null');
-
-        $schema->getType('NonExistingType');
-    }
-
-    public function testFailsOnNonType(): void
-    {
-        $schema = new Schema([
-            'query'      => $this->query,
-            'typeLoader' => static function (): stdClass {
-                return new stdClass();
-            },
-        ]);
-
-        $this->expectException(InvariantViolation::class);
-        $this->expectExceptionMessage('Type loader is expected to return a callable or valid type "Node", but it returned instance of stdClass');
-
-        $schema->getType('Node');
-    }
-
     public function testFailsOnInvalidLoad(): void
     {
         $schema = new Schema([
             'query'      => $this->query,
-            'typeLoader' => function (): Type {
-                return Schema::resolveType($this->content);
-            },
+            'typeLoader' => fn (): Type => Schema::resolveType($this->content),
         ]);
 
+        $expectedType = 'Node';
         $this->expectException(InvariantViolation::class);
-        $this->expectExceptionMessage('Type loader is expected to return type "Node", but it returned "Content"');
+        $this->expectExceptionMessage(Schema::typeLoaderWrongTypeName($expectedType, 'Content'));
 
-        $schema->getType('Node');
-    }
-
-    public function testPassesThroughAnExceptionInLoader(): void
-    {
-        $schema = new Schema([
-            'query'      => $this->query,
-            'typeLoader' => static function (): void {
-                throw new Exception('This is the exception we are looking for');
-            },
-        ]);
-
-        $this->expectException(Throwable::class);
-        $this->expectExceptionMessage('This is the exception we are looking for');
-
-        $schema->getType('Node');
-    }
-
-    public function testReturnsIdenticalResults(): void
-    {
-        $withoutLoader = new Schema([
-            'query'    => $this->query,
-            'mutation' => $this->mutation,
-        ]);
-
-        $withLoader = new Schema([
-            'query'      => $this->query,
-            'mutation'   => $this->mutation,
-            'typeLoader' => $this->typeLoader,
-        ]);
-
-        self::assertSame($withoutLoader->getQueryType(), $withLoader->getQueryType());
-        self::assertSame($withoutLoader->getMutationType(), $withLoader->getMutationType());
-        self::assertSame($withoutLoader->getType('BlogStory'), $withLoader->getType('BlogStory'));
-        self::assertSame($withoutLoader->getDirectives(), $withLoader->getDirectives());
-    }
-
-    public function testSkipsLoaderForInternalTypes(): void
-    {
-        $schema = new Schema([
-            'query'      => $this->query,
-            'mutation'   => $this->mutation,
-            'typeLoader' => $this->typeLoader,
-        ]);
-
-        $type = $schema->getType('ID');
-        self::assertSame(Type::id(), $type);
-        self::assertEquals([], $this->calls);
+        $schema->getType($expectedType);
     }
 }
